+++ src/NPCs/EvilCowboy.java
@@ -18,7 +18,7 @@ public class EvilCowboy extends NPC
         super(id, point.x, point.y, new SpriteSheet(ImageLoader.load("EvilCowboy.png"), 14, 19), "STAND_DOWN");
     }
 
-    public EvilCowboy(int id, Utils.Point point, double startingHealth){
+    public EvilCowboy(int id, Utils.Point point, int startingHealth){
         super(id, point.x, point.y, new SpriteSheet(ImageLoader.load("EvilCowboy.png"), 14, 19), "STAND_DOWN", startingHealth);
     }
